requirements.txt
fastapi
uvicorn
selenium
gspread
oauth2client
google-auth
google-auth-oauthlib
google-auth-httplib2
